Bundle Feature: Create Package Bundle (#677)
What type of PR is this?

Uncomment only one /kind <> line, hit enter to put that in a new line, and remove leading whitespaces from that line:
/component kudoctl
/component operator
/kind feature

What this PR does / why we need it:
This adds 1 feature identified in draft kep-15. It bundles a package into a tarball. It provides a small amount of validation in that the operator and params file must be present. We should add linting at some point. It allows for an output location and opt-in overwrite. It creates bundles based on the expected naming convention which is <operator.name>-<operator-version>.tgz.
It provides error handling around valid destination, valid operator folder and tgz exists.

Which issue(s) this PR fixes:
Fixes #

Special notes for your reviewer:

Does this PR introduce a user-facing change?:

`kudo bundle <operator folder> creates a tgz file for the operator named based on <operator.name>-<operator-version>.tgz`