Use code snippets for code examples
To ensure that examples are in sync with KUDO, their code is taken directly from the KUDO repository. The KUDO repository is added as a submodule. The [markdown-it-vuepress-code-snippet-enhanced](https://github.com/fabbballin/markdown-it-vuepress-code-snippet-enhanced) plugin is used for this, because it supports viewing only a range of lines of a snippet.