Cleanup obsolete examples (#49)
Currently, in 0.7.0 planexecution does not exist so remove that from all the examples referencing it.
Also removed flink and backup example. We should bring those back at one point but currently those were very obsolete and not really working on KUDO for quite some time so we're better without them.